Excluding files/directories starting with "."
Excluding files/directories starting with "." from zip file. If you include the add-on source files for the lib .jar files in zip, directly from the Eclipse project directories it excludes .classpath, .project and .settings. I also have have set .bin as the default output folder in Eclipse. In this way the compiled classes are also excluded from the source files.