Fix #2386 Special characters in add-on translations wrongly displayed
Added lib devtool.jar with method DevtoolUtils.escapeUtf8 which is a
stripped down version of StringEscapeUtils.escapeJava: it only escapes
UTF-8 non-ASCII characters. It escapes them to \uXXXX where XXXX is the
hex code of the escaped character. In releaseAddOn.groovy and
encodeTranslations.groovy StringEscapeUtils.escapeJava is replaced with
DevtoolUtils.escapeUtf8. This partly resolves bug #2386 because it
prevents escaping of backslashes in characters that are already escaped,
like '\n'. The other part is resolved by a commit for Freeplane that
removes StringEscapeUtils.escapeJava from AddOnProperties.java.

The source for devtool.jar, DevtoolUtils.groovy, is added to
devtools/zips/src/addons/devtools. In order to make it part of package
addons.devtools the folder devtools/zips/src/ is added as a source to
the build path. I also made sure that all scripts directories in the
repository are added as a source to the build path.By putting the source
in the zips folder, it is also shipped with the add-on. I added
.settings/org.eclipse.jdt.groovy.core.prefs to .gitignore.