Fix bug #2642 Devtools creates wrong menuTitleKey for scripts
The devtools script `checkAddOn.groovy` adds
`addon.${name}.<scriptname>` as `menuTitleKey` to the script nodes.
According to the note of the translation node this should be
`addons.${name}.<scriptname>`. In addition `checkAddOn.groovy` adds the
add-on name to the translation keys as `addons.${name}`.

To fix this line 440 of checkAddOn.groovy should be changed from:

`def menuTitleKey = "addon.\${name}.${scriptBaseName}"`

to:

`def menuTitleKey = "addons.\${name}.${scriptBaseName}"`

https://sourceforge.net/p/freeplane/bugs/2642/