Fix bug #2643 Devtools does not display its add-on name in menu
All scripts in the devtools add-on have `main_menu_scripting/devtools`
as their `menuLocation`.  Under translations there is also a translation
key `main_menu_scripting/devtools` with the name of the add-on as a
value. For example for locale `en` this is 'Developer Tools'.

However if you install devtools, this key is not translated: the 'Tools'
menu has submenu 'Devtools' and not 'Developer Tools' which you would
expect if you use `en` as a locale. Please note that the translations
for the menu items for the scripts themselves are displayed properly.

I fixed this by simply changing the translation key
`main_menu_scripting/devtools` to `devtools`.

An  alternative fix is that you would change all instances of attribute
`menuLocation` to `main_menu_scripting/addons.${name}`. The submenu
then uses translation key `addons.${name}` directly and there is no need
for an extra translation key `devtools`.

https://sourceforge.net/p/freeplane/bugs/2643/