Bug #2847: Devtools/Gradle plugin incompatibility
https://sourceforge.net/p/freeplane/bugs/2847/
Devtools' checkAddOn.groovy not compatible with
Gradle plugin's directory structure

Devtools' Build add-on function
(checkAddOn.groovy) will now also add script nodes
for scripts that are kept in the Gradle plugin
scripts folder. It does this in addition to the
original functionality: adding script nodes from
the scrips folder relative to the add-on
definition mind map. This means this functionality
will now work when using Devtools standalone and
in combination with the Gradle plugin.

Applied changes

Exclude file name freeplane.gdsl from making
script nodes (like file name freeplane.dsld).

Only look for scripts on the Gradle plugin
location if this plugin is used. In that case
'src' should be the name of the directory one
level higher than the directory of add-on
definition mind map. If the location of this mind
map is customized in build.gradle the name of this
directory might differ. In that case it is not
possible to determine the location of Gradle
plugin scripts and the Build add-on funcion will
fail any way.

Include all file names in sub directories of
src/scripts when making script nodes. The official
directory is src/scripts/groovy. However freeplane
also allows JavaScript and, if you include an
interpreter library, also other scripting
languages. It is not sure whether the Gradle
plugin supports other scripting languages now, but
in this way this fix prepares for the future.